+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where a Neo block type's `conditions` property was not being set in the project config if there were no conditions on the block type
+
 ## 3.5.10 - 2022-12-06
 
 ### Fixed